requirements2.txt
black==22.3.0
checksumdir
google-api-python-client==2.51.0
google-cloud-firestore==2.5.3
google-cloud-pubsub==2.13.0
google-cloud-storage==2.4.0
ipaddr==2.2.0
PyNaCl==1.5.0
tenacity==8.0.1
tqdm==4.64.0
